Add a simple benchmark of two implementations of a word count in Go
This approach uses the shell's time function and a fixed number of iterations.
The executable requires specifying which implementation to use via a command line argument.

The intent of this repository is to gradually change the source code to benchmark in a
way that is more consistent with idiomatic Go benchmarking.